+++ scripts/export-mrpack.sh
@@ -9,5 +9,5 @@ gtl_output="$gtl_output_dir/GregTech-Leisure-$gtl_version.mrpack"
 mkdir -p "$gtl_output_dir"
 cd "$gtl_root"
 "$gtl_root/scripts/packwiz-command.sh" refresh
-"$gtl_root/scripts/packwiz-command.sh" modrinth export -o "$gtl_output"
+"$gtl_root/scripts/packwiz-command.sh" modrinth export --restrictDomains=false -o "$gtl_output"
 shasum -a 256 "$gtl_output"